+++ src/runtime/bddStepInvoker.ts
@@ -7,7 +7,7 @@ import { getLocationInFile } from '../playwright/getLocationInFile';
 import { DataTable } from '../cucumber/DataTable';
 import { BddAutoInjectFixtures, getBddAutoInjectFixtures } from './bddTestFixturesAuto';
 import { runStepWithLocation } from '../playwright/runStepWithLocation';
-import { formatDuplicateStepsMessage, StepFinder } from '../steps/finder';
+import { excludeStepAliases, formatDuplicateStepsMessage, StepFinder } from '../steps/finder';
 import { MatchedStepDefinition } from '../steps/matchedStepDefinition';
 import { BddContext } from './bddContext';
 import { getStepHooksToRun, runStepHooks } from '../hooks/step';
@@ -113,14 +113,15 @@ export class BddStepInvoker {
 
   private findStepDefinition(stepText: string, stepTextWithKeyword: string) {
     const { keywordType, gherkinStepLine } = this.getBddStepData();
-    const stepDefinitions = this.stepFinder.findDefinitions(
+    let stepDefinitions = this.stepFinder.findDefinitions(
       keywordType,
       stepText,
       this.bddContext.tags,
     );
+    stepDefinitions = excludeStepAliases(stepDefinitions);
 
-    const firstFoundDefinition = stepDefinitions[0];
-    if (firstFoundDefinition && stepDefinitions.length === 1) return firstFoundDefinition;
+    const firstDefinition = stepDefinitions[0];
+    if (stepDefinitions.length === 1 && firstDefinition) return firstDefinition;
 
     const fullStepLocation = `${this.bddContext.featureUri}:${gherkinStepLine}`;