+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format.
 
 ## [Unreleased]
+* feat: support multiple aliases for step definitions
 
 ## [9.2.0] - 2026-06-18
 * feat: include CI information in Cucumber reports